requirements.txt
Flask==2.0.2
psycopg2==2.8.5
uWSGI==2.0.20
APScheduler==3.8.1
python-dateutil==2.8.2